requirements.txt
# Core AI/LLM
langchain==0.1.0
langchain-openai==0.0.5
langchain-aws==0.1.0  # For AWS Bedrock (Claude)
langchain-anthropic==0.1.1  # For Anthropic Claude
openai==1.12.0
langchain-community==0.0.20
chromadb==0.4.22
sentence-transformers==2.3.1

# AWS Integration
boto3==1.34.51
botocore==1.34.51

# Web Automation & Screenshots
selenium==4.17.2
undetected-chromedriver==3.5.5  # Anti-detection Selenium for Cisco Duo
playwright==1.41.2
pillow==10.2.0
opencv-python==4.8.1.78  # For intelligent image analysis
numpy==1.26.2  # For image array operations
pyautogui==0.9.54
mss==9.0.1

# API Integrations
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.3

# Microsoft/SharePoint
Office365-REST-Python-Client==2.5.3
msal==1.26.0
msgraph-sdk==1.1.0

# Webex
webexteamssdk==1.6.1

# PagerDuty
pdpyras==5.2.0

# Datadog
datadog-api-client==2.20.0

# Splunk
splunk-sdk==1.7.4

# Kubernetes
kubernetes==29.0.0

# Elasticsearch/Kibana
elasticsearch==8.12.0

# Data Processing & Export
pandas==2.2.0
openpyxl==3.1.2
xlsxwriter==3.1.9
fpdf2==2.7.7
reportlab==4.1.0
weasyprint==60.2

# Document Processing & OCR
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==1.1.0  # For Word document generation

# Configuration & Secrets
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
pyyaml==6.0.1
keyring==24.3.0

# CLI & UI
typer==0.9.0
rich==13.7.0
questionary==2.0.1
prompt-toolkit==3.0.43

# Utilities
pytz==2024.1
python-dateutil==2.8.2
tenacity==8.2.3

# Logging
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4

playwright>=1.40.0
